[CHORE]: testdi -> di
