+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:required="false" />
 
     <application
-        android:name=".testdi.PlayTogetherApplication"
+        android:name=".di.PlayTogetherApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
