+++ app/src/main/java/com/playtogether_android/app/di/PlayTogetherApplication.kt
@@ -1,27 +1,35 @@
-//package com.playtogether_android.app.di
-//
-//import android.app.Application
-//import android.content.Context
-//import androidx.databinding.ktx.BuildConfig
-//import org.koin.android.ext.koin.androidContext
-//import org.koin.android.ext.koin.androidLogger
-//import org.koin.core.context.startKoin
-//import org.koin.core.logger.Level
-//
-//class PlayTogetherApplication : Application() {
-//    init {
-//        instance = this
-//    }
-//
-//    companion object {
-//        var instance: PlayTogetherApplication? = null
-//        fun context(): Context {
-//            return instance!!.applicationContext
-//        }
-//    }
-//
-//    override fun onCreate() {
-//        super.onCreate()
+package com.playtogether_android.app.di
+
+import android.app.Activity
+import android.app.Application
+import android.content.Context
+import android.os.Bundle
+import com.playtogether_android.app.util.PixelRatio
+import com.playtogether_android.data.singleton.PlayTogetherRepository
+import dagger.hilt.android.HiltAndroidApp
+import timber.log.Timber
+
+@HiltAndroidApp
+class PlayTogetherApplication : Application(), Application.ActivityLifecycleCallbacks {
+    init {
+        instance = this
+    }
+
+    companion object {
+        var instance: PlayTogetherApplication? = null
+        fun context(): Context {
+            return instance!!.applicationContext
+        }
+
+        lateinit var pixelRatio: PixelRatio
+        var isBackground = false
+    }
+
+    override fun onCreate() {
+        super.onCreate()
+        registerActivityLifecycleCallbacks(this)
+        initLogger()
+        PlayTogetherRepository.init(this)
 //        startKoin {
 //            androidLogger(if (BuildConfig.DEBUG) Level.ERROR else Level.NONE)
 //            androidContext(this@PlayTogetherApplication)
@@ -32,5 +40,43 @@
 //            modules(useCaseModule)
 //            modules(apiModule)
 //        }
-//    }
-//}
+    }
+
+    override fun onTerminate() {
+        super.onTerminate()
+        unregisterActivityLifecycleCallbacks(this)
+    }
+
+    private fun initLogger() {
+        Timber.plant(Timber.DebugTree())
+    }
+
+    override fun onActivityCreated(p0: Activity, p1: Bundle?) {
+
+    }
+
+    override fun onActivityStarted(p0: Activity) {
+        isBackground = false
+    }
+
+    override fun onActivityResumed(p0: Activity) {
+        isBackground = false
+    }
+
+    override fun onActivityPaused(p0: Activity) {
+        isBackground = true
+    }
+
+    override fun onActivityStopped(p0: Activity) {
+
+    }
+
+    override fun onActivitySaveInstanceState(p0: Activity, p1: Bundle) {
+
+    }
+
+    override fun onActivityDestroyed(p0: Activity) {
+
+    }
+}
+
